config/urls.py
"""config URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.0/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path
from ninja import NinjaAPI

from config.utils.permissions import AuthBearer
from account.views import auth_controller
from healthcare.views.doctor import doctor_controller
from healthcare.views.diagnosis_request import diagnosis_request_controller

api = NinjaAPI(
    version="1.0.0",
    title="healthcare api",
    description="healthcare app api docs",
    auth=AuthBearer()
)

api.add_router("auth", auth_controller)
api.add_router("doctor", doctor_controller)
api.add_router("diagnosis-request", diagnosis_request_controller)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', api.urls)
]
