pyproject.toml
[tool.poetry]
name = "healthcare"
version = "0.1.0"
description = "sample healthcare app based on django & django-ninja"
authors = ["Sean Kim <ddhyun93@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.9"
django-ninja = "^0.19.1"
mysqlclient = "^2.1.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
